+++ src/app/history/partials/List.tsx
@@ -1,8 +1,8 @@
 'use client';
 import React, { useEffect, useState } from 'react';
-import Card from '@/components/Card';
 import useLocalStorage from '@/hooks/useLocalStorage';
 import { CompactManga } from '@/types';
+import { MangaCard } from '@/components/manga-card';
 export default function List() {
   const { getWatched, delWatched } = useLocalStorage();
   const [mangas, setmangas] = useState<any | null>(null);
@@ -14,7 +14,7 @@ export default function List() {
   return mangas !== null && Object.keys(mangas).length > 0 ? (
     <>
       {mangas.map((manga: CompactManga) => (
-        <Card manga={manga} />
+        <MangaCard manga={manga} />
       ))}
     </>
   ) : (
