+++ src/app/page.tsx
@@ -1,5 +1,5 @@
-import Carousel from '@/components/Carousel';
-import Card from '@/components/manga-card';
+import { Carousel } from '@/components/carousel';
+import { MangaCard } from '@/components/manga-card';
 import { useBackend } from '@/hooks/useBackend';
 
 export default async function Home() {
@@ -17,13 +17,13 @@ export default async function Home() {
 
           <div className='grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4 lg:gap-6 mb-10 px-3'>
             {feed.map((manga: any) => (
-              <Card manga={manga} key={manga.id} />
+              <MangaCard manga={manga} key={manga.id} />
             ))}
           </div>
 
           {/* <div className='flex flex-wrap justify-center gap-3 bigp:gap-6 bigp:mx-24'>
             {feed.map((manga: any, index: number) => (
-              <Card manga={manga} key={manga.id} />
+              <MangaCard manga={manga} key={manga.id} />
             ))}
           </div> */}
           {/* <p
@@ -33,7 +33,7 @@ export default async function Home() {
           </p>
           <div className='flex flex-wrap justify-center gap-3 bigp:mx-24 bigp:gap-6'>
             {feed.map((manga: any, index: number) => (
-              <Card
+              <MangaCard
                 id={manga.id}
                 key={index}
                 title={manga.title.romaji}
