+++ src/components/ChapterList.tsx
@@ -1,8 +1,10 @@
 'use client';
+
 import React, { useEffect, useState } from 'react';
 import useManga from '@/hooks/useManga';
 import { useRouter, useParams } from 'next/navigation';
-import Loading from '@/components/LoadingSingle';
+import Loading from '@/components/loading-single';
+
 export default async function ChapterList() {
   const { getChapters } = useManga();
   const anilistId = useParams().slug;
